Verified 'Must do changes' implemented all 'should do' changes
Must do Changes that have been verified:
-Change the Clock divider chip to the 2,5V part
-Remove AC coupling caps in clock lines

Should do changes implmented
  - Add pull up resistor to ADC reset lines
  - Internal termination of PPS (when unused)
  - Add Heat Sinking to inductors by increased traces/planes size (Copied/Based-On from V2 Design)
  - Changed to a 2.048V VREF with accompanying removal of INA851 gain resistor (this will give more dynamic range)
  - Move ground test points further away from power rails

Also Included Some Improvments to the power rails seen in V2 when in comes to:
-12V rail into the power section and up to the front of the board
-The 3.3V rail that goes around the edge of the board to the relay switch IC's

Note - DRC warnings that exist are either ignorable, or pointing to the fact that the default footprint libs may have been changed in Kicad9 causeing a mismatch to what is on the board. I will no update these footprint to the new version yet, or possibly ever.